Extract IntentFirstConfirmationInterceptor.kt
