+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/intent/IntentConfirmationInterceptor.kt
@@ -1,26 +1,22 @@
 package com.stripe.android.paymentelement.confirmation.intent
 
-import com.stripe.android.ConfirmStripeIntentParamsFactory
 import com.stripe.android.SharedPaymentTokenSessionPreview
 import com.stripe.android.core.exception.StripeException
 import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
 import com.stripe.android.core.networking.ApiRequest
-import com.stripe.android.core.strings.resolvableString
 import com.stripe.android.model.ConfirmPaymentIntentParams
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.PaymentMethodExtraParams
 import com.stripe.android.model.PaymentMethodOptionsParams
-import com.stripe.android.model.RadarOptions
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.StripeIntent
 import com.stripe.android.networking.StripeRepository
 import com.stripe.android.paymentelement.PreparePaymentMethodHandler
 import com.stripe.android.paymentelement.confirmation.ALLOWS_MANUAL_CONFIRMATION
 import com.stripe.android.paymentelement.confirmation.ConfirmationDefinition
-import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
 import com.stripe.android.paymentelement.confirmation.intent.IntentConfirmationDefinition.Args
 import com.stripe.android.payments.core.analytics.ErrorReporter
 import com.stripe.android.paymentsheet.CreateIntentCallback
@@ -29,7 +25,6 @@ import com.stripe.android.paymentsheet.state.PaymentElementLoader
 import javax.inject.Inject
 import javax.inject.Named
 import javax.inject.Provider
-import com.stripe.android.R as PaymentsCoreR
 
 internal interface IntentConfirmationInterceptor {
 
@@ -163,23 +158,17 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
                 }
             }
 
-            is PaymentElementLoader.InitializationMode.PaymentIntent -> {
-                createConfirmStep(
-                    clientSecret = initializationMode.clientSecret,
+            is PaymentElementLoader.InitializationMode.PaymentIntent,
+            is PaymentElementLoader.InitializationMode.SetupIntent -> {
+                IntentFirstConfirmationInterceptor(
+                    publishableKeyProvider = publishableKeyProvider,
+                    stripeAccountIdProvider = stripeAccountIdProvider,
+                ).handleNewPaymentMethod(
+                    initializationMode = initializationMode,
                     intent = intent,
                     shippingValues = shippingValues,
                     paymentMethodCreateParams = paymentMethodCreateParams,
                     paymentMethodOptionsParams = paymentMethodOptionsParams,
-                    paymentMethodExtraParams = paymentMethodExtraParams,
-                )
-            }
-
-            is PaymentElementLoader.InitializationMode.SetupIntent -> {
-                createConfirmStep(
-                    clientSecret = initializationMode.clientSecret,
-                    intent = intent,
-                    shippingValues = shippingValues,
-                    paymentMethodCreateParams = paymentMethodCreateParams,
                     paymentMethodExtraParams = paymentMethodExtraParams,
                 )
             }
@@ -231,23 +220,13 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
                 }
             }
 
-            is PaymentElementLoader.InitializationMode.PaymentIntent -> {
-                createConfirmStep(
-                    clientSecret = initializationMode.clientSecret,
-                    intent = intent,
-                    shippingValues = shippingValues,
-                    paymentMethod = paymentMethod,
-                    paymentMethodOptionsParams = paymentMethodOptionsParams,
-                    paymentMethodExtraParams = paymentMethodExtraParams,
-                    isDeferred = false,
-                    intentConfigSetupFutureUsage = null,
-                    hCaptchaToken = hCaptchaToken
-                )
-            }
-
+            is PaymentElementLoader.InitializationMode.PaymentIntent,
             is PaymentElementLoader.InitializationMode.SetupIntent -> {
-                createConfirmStep(
-                    clientSecret = initializationMode.clientSecret,
+                IntentFirstConfirmationInterceptor(
+                    publishableKeyProvider = publishableKeyProvider,
+                    stripeAccountIdProvider = stripeAccountIdProvider,
+                ).handleSavedPaymentMethod(
+                    initializationMode = initializationMode,
                     intent = intent,
                     shippingValues = shippingValues,
                     paymentMethod = paymentMethod,
@@ -261,85 +240,6 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
         }
     }
 
-    private fun createConfirmStep(
-        clientSecret: String,
-        intent: StripeIntent,
-        shippingValues: ConfirmPaymentIntentParams.Shipping?,
-        paymentMethod: PaymentMethod,
-        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
-        paymentMethodExtraParams: PaymentMethodExtraParams?,
-        isDeferred: Boolean,
-        intentConfigSetupFutureUsage: ConfirmPaymentIntentParams.SetupFutureUsage?,
-        hCaptchaToken: String?
-    ): ConfirmationDefinition.Action<Args> {
-        val factory = ConfirmStripeIntentParamsFactory.createFactory(
-            clientSecret = clientSecret,
-            intent = intent,
-            shipping = shippingValues,
-        ) ?: run {
-            val exception = InvalidClientSecretException(clientSecret, intent)
-
-            return createFailStep(exception, exception.message)
-        }
-
-        val confirmParams = factory.create(
-            paymentMethod = paymentMethod,
-            optionsParams = paymentMethodOptionsParams,
-            extraParams = paymentMethodExtraParams,
-            intentConfigSetupFutureUsage = intentConfigSetupFutureUsage,
-            radarOptions = hCaptchaToken?.let { RadarOptions(it) }
-        )
-        return ConfirmationDefinition.Action.Launch(
-            launcherArguments = Args.Confirm(confirmParams),
-            deferredIntentConfirmationType = DeferredIntentConfirmationType.Client.takeIf { isDeferred },
-            receivesResultInProcess = false,
-        )
-    }
-    private fun createConfirmStep(
-        clientSecret: String,
-        intent: StripeIntent,
-        shippingValues: ConfirmPaymentIntentParams.Shipping?,
-        paymentMethodCreateParams: PaymentMethodCreateParams,
-        paymentMethodOptionsParams: PaymentMethodOptionsParams? = null,
-        paymentMethodExtraParams: PaymentMethodExtraParams? = null,
-    ): ConfirmationDefinition.Action<Args> {
-        val paramsFactory = ConfirmStripeIntentParamsFactory.createFactory(
-            clientSecret = clientSecret,
-            intent = intent,
-            shipping = shippingValues,
-        ) ?: run {
-            val exception = InvalidClientSecretException(clientSecret, intent)
-
-            return createFailStep(exception, exception.message)
-        }
-
-        val confirmParams = paramsFactory.create(
-            paymentMethodCreateParams,
-            paymentMethodOptionsParams,
-            paymentMethodExtraParams,
-        )
-
-        return ConfirmationDefinition.Action.Launch(
-            launcherArguments = Args.Confirm(confirmParams),
-            deferredIntentConfirmationType = null,
-            receivesResultInProcess = false,
-        )
-    }
-
-    private fun createFailStep(
-        exception: Exception,
-        message: String,
-    ): ConfirmationDefinition.Action<Args> {
-        return ConfirmationDefinition.Action.Fail(
-            cause = exception,
-            message = if (requestOptions.apiKeyIsLiveMode) {
-                PaymentsCoreR.string.stripe_internal_error.resolvableString
-            } else {
-                message.resolvableString
-            },
-            errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
-        )
-    }
     internal companion object {
         const val PROVIDER_FETCH_TIMEOUT = 2
         const val PROVIDER_FETCH_INTERVAL = 5L
