+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/intent/IntentFirstConfirmationInterceptor.kt
@@ -1,0 +1,153 @@
+package com.stripe.android.paymentelement.confirmation.intent
+
+import com.stripe.android.ConfirmStripeIntentParamsFactory
+import com.stripe.android.core.injection.PUBLISHABLE_KEY
+import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
+import com.stripe.android.core.networking.ApiRequest
+import com.stripe.android.core.strings.resolvableString
+import com.stripe.android.model.ConfirmPaymentIntentParams
+import com.stripe.android.model.PaymentMethod
+import com.stripe.android.model.PaymentMethodCreateParams
+import com.stripe.android.model.PaymentMethodExtraParams
+import com.stripe.android.model.PaymentMethodOptionsParams
+import com.stripe.android.model.RadarOptions
+import com.stripe.android.model.StripeIntent
+import com.stripe.android.paymentelement.confirmation.ConfirmationDefinition
+import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
+import com.stripe.android.paymentelement.confirmation.intent.IntentConfirmationDefinition.Args
+import com.stripe.android.paymentsheet.state.PaymentElementLoader
+import javax.inject.Named
+import com.stripe.android.R as PaymentsCoreR
+
+class IntentFirstConfirmationInterceptor(
+    @Named(PUBLISHABLE_KEY) private val publishableKeyProvider: () -> String,
+    @Named(STRIPE_ACCOUNT_ID) private val stripeAccountIdProvider: () -> String?,
+) {
+
+    private val requestOptions: ApiRequest.Options
+        get() = ApiRequest.Options(
+            apiKey = publishableKeyProvider(),
+            stripeAccount = stripeAccountIdProvider(),
+        )
+    internal fun handleNewPaymentMethod(
+        initializationMode: PaymentElementLoader.IntentFirst,
+        intent: StripeIntent,
+        paymentMethodCreateParams: PaymentMethodCreateParams,
+        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
+        paymentMethodExtraParams: PaymentMethodExtraParams?,
+        shippingValues: ConfirmPaymentIntentParams.Shipping?,
+    ): ConfirmationDefinition.Action<Args> {
+        return createConfirmStep(
+            clientSecret = initializationMode.clientSecret,
+            intent = intent,
+            shippingValues = shippingValues,
+            paymentMethodCreateParams = paymentMethodCreateParams,
+            paymentMethodOptionsParams = paymentMethodOptionsParams,
+            paymentMethodExtraParams = paymentMethodExtraParams,
+        )
+    }
+
+    internal fun handleSavedPaymentMethod(
+        initializationMode: PaymentElementLoader.IntentFirst,
+        intent: StripeIntent,
+        paymentMethod: PaymentMethod,
+        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
+        paymentMethodExtraParams: PaymentMethodExtraParams?,
+        shippingValues: ConfirmPaymentIntentParams.Shipping?,
+        isDeferred: Boolean,
+        intentConfigSetupFutureUsage: ConfirmPaymentIntentParams.SetupFutureUsage?,
+        hCaptchaToken: String?,
+    ): ConfirmationDefinition.Action<Args> {
+        return createConfirmStep(
+            clientSecret = initializationMode.clientSecret,
+            intent = intent,
+            shippingValues = shippingValues,
+            paymentMethod = paymentMethod,
+            paymentMethodOptionsParams = paymentMethodOptionsParams,
+            paymentMethodExtraParams = paymentMethodExtraParams,
+            isDeferred = isDeferred,
+            intentConfigSetupFutureUsage = intentConfigSetupFutureUsage,
+            hCaptchaToken = hCaptchaToken,
+        )
+    }
+    private fun createConfirmStep(
+        clientSecret: String,
+        intent: StripeIntent,
+        shippingValues: ConfirmPaymentIntentParams.Shipping?,
+        paymentMethod: PaymentMethod,
+        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
+        paymentMethodExtraParams: PaymentMethodExtraParams?,
+        isDeferred: Boolean,
+        intentConfigSetupFutureUsage: ConfirmPaymentIntentParams.SetupFutureUsage?,
+        hCaptchaToken: String?
+    ): ConfirmationDefinition.Action<Args> {
+        val factory = ConfirmStripeIntentParamsFactory.createFactory(
+            clientSecret = clientSecret,
+            intent = intent,
+            shipping = shippingValues,
+        ) ?: run {
+            val exception = InvalidClientSecretException(clientSecret, intent)
+
+            return createFailStep(exception, exception.message)
+        }
+
+        val confirmParams = factory.create(
+            paymentMethod = paymentMethod,
+            optionsParams = paymentMethodOptionsParams,
+            extraParams = paymentMethodExtraParams,
+            intentConfigSetupFutureUsage = intentConfigSetupFutureUsage,
+            radarOptions = hCaptchaToken?.let { RadarOptions(it) }
+        )
+        return ConfirmationDefinition.Action.Launch(
+            launcherArguments = Args.Confirm(confirmParams),
+            deferredIntentConfirmationType = DeferredIntentConfirmationType.Client.takeIf { isDeferred },
+            receivesResultInProcess = false,
+        )
+    }
+
+    private fun createConfirmStep(
+        clientSecret: String,
+        intent: StripeIntent,
+        shippingValues: ConfirmPaymentIntentParams.Shipping?,
+        paymentMethodCreateParams: PaymentMethodCreateParams,
+        paymentMethodOptionsParams: PaymentMethodOptionsParams? = null,
+        paymentMethodExtraParams: PaymentMethodExtraParams? = null,
+    ): ConfirmationDefinition.Action<Args> {
+        val paramsFactory = ConfirmStripeIntentParamsFactory.createFactory(
+            clientSecret = clientSecret,
+            intent = intent,
+            shipping = shippingValues,
+        ) ?: run {
+            val exception = InvalidClientSecretException(clientSecret, intent)
+
+            return createFailStep(exception, exception.message)
+        }
+
+        val confirmParams = paramsFactory.create(
+            paymentMethodCreateParams,
+            paymentMethodOptionsParams,
+            paymentMethodExtraParams,
+        )
+
+        return ConfirmationDefinition.Action.Launch(
+            launcherArguments = Args.Confirm(confirmParams),
+            deferredIntentConfirmationType = null,
+            receivesResultInProcess = false,
+        )
+    }
+
+    private fun createFailStep(
+        exception: Exception,
+        message: String,
+    ): ConfirmationDefinition.Action<Args> {
+        return ConfirmationDefinition.Action.Fail(
+            cause = exception,
+            message = if (requestOptions.apiKeyIsLiveMode) {
+                PaymentsCoreR.string.stripe_internal_error.resolvableString
+            } else {
+                message.resolvableString
+            },
+            errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
+        )
+    }
+}
